+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [#344](https://github.com/nf-core/airrflow/pull/344) Added option to save non-productive sequences
 - [#344](https://github.com/nf-core/airrflow/pull/344) Allow for maskprimers align with UMI on both V gene and C gene primers.
 - [#344](https://github.com/nf-core/airrflow/pull/344) Update pipeline template to nf-core tools v3.0.2.
+- Added support for AIRR community germline reference data.
 
 ### `Fixed`
 
@@ -25,6 +26,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### `Dependencies`
 
 ### `Deprecated parameters`
+
+- `--fetch_imgt` is now deprecated in favor of `--fetch_reference`. This new parameter is false by default, using the cached `imgt` reference data. When set to `imgt` or `airrc` it will download the latest version of the [IMGT](https://www.imgt.org/) or [AIRR community](https://ogrdb.airr-community.org/).
+
 
 ## [4.1.0] - Avenseguim